input: merge to libtrx
