+++ src/tr1/game/gun/gun_misc.c
@@ -1,7 +1,6 @@
 #include "game/gun/gun_misc.h"
 
 #include "game/game.h"
-#include "game/input.h"
 #include "game/inventory.h"
 #include "game/los.h"
 #include "game/random.h"
@@ -14,6 +13,7 @@
 #include <libtrx/config.h>
 #include <libtrx/game/collision.h>
 #include <libtrx/game/gun/vars.h>
+#include <libtrx/game/input.h>
 #include <libtrx/game/math.h>
 #include <libtrx/game/matrix.h>
 #include <libtrx/utils.h>
