+++ src/tr1/game/gun/gun_rifle.c
@@ -2,7 +2,6 @@
 
 #include "game/gun.h"
 #include "game/gun/gun_misc.h"
-#include "game/input.h"
 #include "game/random.h"
 #include "game/sound.h"
 #include "global/vars.h"
@@ -10,6 +9,7 @@
 #include <libtrx/config.h>
 #include <libtrx/game/camera.h>
 #include <libtrx/game/gun/vars.h>
+#include <libtrx/game/input.h>
 
 #include <stdint.h>
