+++ src/tr1/game/objects/general/scion1.c
@@ -2,7 +2,6 @@
 // Triggers O_LARA_EXTRA pedestal pickup animation.
 
 #include "game/game.h"
-#include "game/input.h"
 #include "game/inventory.h"
 #include "game/lara.h"
 #include "game/level.h"
@@ -12,6 +11,7 @@
 #include "global/vars.h"
 
 #include <libtrx/game/camera.h>
+#include <libtrx/game/input.h>
 #include <libtrx/game/overlay.h>
 
 #define EXTRA_ANIM_PEDESTAL_SCION 0
