+++ src/tr1/game/objects/general/scion4.c
@@ -1,10 +1,10 @@
 // Atlantis Scion - triggers O_LARA_EXTRA reach anim.
 
-#include "game/input.h"
 #include "game/lara.h"
 #include "global/vars.h"
 
 #include <libtrx/game/camera.h>
+#include <libtrx/game/input.h>
 
 #define EXTRA_ANIM_HOLDER_SCION 0
