+++ src/tr1/game/objects/general/switch.c
@@ -1,9 +1,9 @@
-#include "game/input.h"
 #include "game/lara.h"
 #include "game/objects/common.h"
 #include "global/vars.h"
 
 #include <libtrx/config.h>
+#include <libtrx/game/input.h>
 
 typedef enum {
     SWITCH_STATE_OFF = 0,
