+++ src/tr1/game/option/option_compass.c
@@ -3,10 +3,10 @@
 #include "game/game.h"
 #include "game/game_flow.h"
 #include "game/game_string.h"
-#include "game/input.h"
 #include "global/vars.h"
 
 #include <libtrx/config.h>
+#include <libtrx/game/input.h>
 #include <libtrx/game/ui.h>
 
 #include <stdint.h>
