+++ src/tr1/game/option/option_passport.c
@@ -3,7 +3,6 @@
 #include "game/game.h"
 #include "game/game_flow.h"
 #include "game/game_string.h"
-#include "game/input.h"
 #include "game/inventory.h"
 #include "game/inventory_ring.h"
 #include "game/savegame.h"
@@ -11,6 +10,7 @@
 #include "global/vars.h"
 
 #include <libtrx/config.h>
+#include <libtrx/game/input.h>
 #include <libtrx/game/overlay.h>
 #include <libtrx/game/ui.h>
 #include <libtrx/memory.h>
