+++ src/tr1/meson.build
@@ -117,7 +117,6 @@ sources = [
   'game/gun/gun_misc.c',
   'game/gun/gun_pistols.c',
   'game/gun/gun_rifle.c',
-  'game/input.c',
   'game/inventory.c',
   'game/inventory_ring/control.c',
   'game/inventory_ring/draw.c',
