--- src/tr2/game/input.h
@@ -1,3 +1,0 @@
-#pragma once
-
-#include <libtrx/game/input.h>
