+++ src/tr2/game/inventory_ring/control.c
@@ -3,7 +3,6 @@
 #include "game/demo.h"
 #include "game/game.h"
 #include "game/game_flow.h"
-#include "game/input.h"
 #include "game/inventory.h"
 #include "game/inventory_ring/draw.h"
 #include "game/inventory_ring/vars.h"
@@ -19,6 +18,7 @@
 #include <libtrx/game/clock.h>
 #include <libtrx/game/gun/const.h>
 #include <libtrx/game/gym.h>
+#include <libtrx/game/input.h>
 #include <libtrx/game/interpolation.h>
 #include <libtrx/game/inventory_ring/priv.h>
 #include <libtrx/game/matrix.h>
