+++ src/tr2/game/objects/creatures/dragon.c
@@ -1,5 +1,4 @@
 #include "game/creature.h"
-#include "game/input.h"
 #include "game/lara.h"
 #include "game/objects/common.h"
 #include "game/random.h"
@@ -12,6 +11,7 @@
 #include <libtrx/game/camera.h>
 #include <libtrx/game/carrier.h>
 #include <libtrx/game/collision.h>
+#include <libtrx/game/input.h>
 #include <libtrx/game/lara/common.h>
 #include <libtrx/game/math.h>
 #include <libtrx/game/output.h>
