+++ src/tr2/game/objects/vehicles/boat.c
@@ -1,6 +1,5 @@
 #include "decomp/decomp.h"
 #include "game/effects.h"
-#include "game/input.h"
 #include "game/objects/common.h"
 #include "game/objects/traps/gondola.h"
 #include "game/random.h"
@@ -12,6 +11,7 @@
 #include <libtrx/game/camera.h>
 #include <libtrx/game/collision.h>
 #include <libtrx/game/game_buf.h>
+#include <libtrx/game/input.h>
 #include <libtrx/game/lara.h>
 #include <libtrx/game/math.h>
 #include <libtrx/game/matrix.h>
