+++ src/tr2/game/option/option_compass.c
@@ -1,10 +1,10 @@
 #include "game/game.h"
-#include "game/input.h"
 #include "game/option/option.h"
 #include "game/savegame.h"
 #include "game/sound.h"
 #include "global/vars.h"
 
+#include <libtrx/game/input.h>
 #include <libtrx/game/ui.h>
 
 #include <stdio.h>
