+++ src/tr2/game/option/option_passport.c
@@ -2,7 +2,6 @@
 #include "game/game.h"
 #include "game/game_flow.h"
 #include "game/game_string.h"
-#include "game/input.h"
 #include "game/inventory.h"
 #include "game/inventory_ring.h"
 #include "game/option/option.h"
@@ -12,6 +11,7 @@
 
 #include <libtrx/config.h>
 #include <libtrx/debug.h>
+#include <libtrx/game/input.h>
 #include <libtrx/game/overlay.h>
 #include <libtrx/game/ui.h>
